refactor(base-utils): Rename `evm-utils` to `evm`
