+++ apps/docs.blocksense.network/src/_mock/types.ts
@@ -1,6 +1,6 @@
 import * as S from '@effect/schema/Schema';
 
-import { ethereumAddress, networkName } from '@blocksense/base-utils/evm-utils';
+import { ethereumAddress, networkName } from '@blocksense/base-utils/evm';
 
 // Schema for Chainlink Proxy entries
 const ChainlinkProxySchema = S.Struct({
