+++ libs/base-utils/src/evm/hex-types.ts
