+++ libs/base-utils/src/evm/networks.spec.ts
