+++ libs/base-utils/src/evm/networks.ts
