+++ libs/base-utils/src/index.ts
@@ -3,7 +3,7 @@ export * from './assert';
 export * from './async';
 export * from './buffer-and-hex';
 export * from './env';
-export * from './evm-utils';
+export * from './evm';
 export * from './fs';
 export * from './numeric';
 export * from './string';
